+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## FUTURE
 
+* [ADDED] - `ResourceManager.prototype.tag` method which allows to tag resources and inspect a group of resources via the tag
 * [ADDED] - set project's `version` option on `Service` config object when `Service` is being instantiated
 * [CHANGED] - deprecate `App.prototype.useSession` method